@@ -1,15 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-  compileSdkVersion 28
-  buildToolsVersion "28.0.3"
+  compileSdk 33
   defaultConfig {
     applicationId "com.linchaolong.android.imagepicker.demo"
     minSdkVersion 14
-    targetSdkVersion 28
+    targetSdkVersion 33
     versionCode 1
     versionName "1.0"
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
   }
   buildTypes {
     release {
@@ -20,13 +19,13 @@ android {
 }
 dependencies {
   implementation fileTree(dir: 'libs', include: ['*.jar'])
-  androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
-    exclude group: 'com.android.support', module: 'support-annotations'
-  })
-  implementation 'com.android.support:appcompat-v7:28.0.0'
-  testImplementation 'junit:junit:4.12'
+  implementation 'androidx.appcompat:appcompat:1.0.0'
   implementation 'com.facebook.fresco:fresco:1.1.0'
-  implementation 'com.linchaolong.android:imagepicker:1.5'
-//  compile project(path: ':library')
+//  implementation 'com.linchaolong.android:imagepicker:1.5'
+  implementation project(path: ':library')
+
+  testImplementation "junit:junit:4.13.2"
+  androidTestImplementation "androidx.test.ext:junit:1.1.5"
+  androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
 }
 
