@@ -4,19 +4,17 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AlertDialog;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AlertDialog;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import com.facebook.drawee.generic.RoundingParams;
-import com.facebook.drawee.view.SimpleDraweeView;
+
 import com.linchaolong.android.imagepicker.ImagePicker;
-import com.linchaolong.android.imagepicker.cropper.CropImage;
 import com.linchaolong.android.imagepicker.cropper.CropImageView;
 
 /**
@@ -59,7 +57,7 @@ public class TestFragment extends Fragment {
 
   private void startCameraOrGallery() {
     new AlertDialog.Builder(getActivity()).setTitle("设置头像")
-        .setItems(new String[] { "从相册中选取图片", "拍照" }, new DialogInterface.OnClickListener() {
+        .setItems(new String[] { "选择器(Chooser)", "从相册中选取图片(Gallery)", "拍照(Camera)" }, new DialogInterface.OnClickListener() {
           @Override public void onClick(DialogInterface dialog, int which) {
             // 回调
             ImagePicker.Callback callback = new ImagePicker.Callback() {
@@ -72,6 +70,8 @@ public class TestFragment extends Fragment {
               }
             };
             if (which == 0) {
+              imagePicker.startChooser(TestFragment.this, callback);
+            } else if (which == 1) {
               // 从相册中选取图片
               imagePicker.startGallery(TestFragment.this, callback);
             } else {
