@@ -1,8 +1,8 @@
 package com.linchaolong.android.imagepicker;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -17,7 +17,7 @@ import static org.junit.Assert.*;
 @RunWith(AndroidJUnit4.class) public class ExampleInstrumentedTest {
   @Test public void useAppContext() throws Exception {
     // Context of the app under test.
-    Context appContext = InstrumentationRegistry.getTargetContext();
+    Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
     assertEquals("com.linchaolong.android.imagepicker.test", appContext.getPackageName());
   }
